controller: Move controller to own repository
The `udrone` package is installed on drones which run actions against a
test device. It should be as small and simple as possible and most logic
handled by the controller application, managing active drones.

The controller is under heavy development and is a single Python
package. It does not make sense to mix a C compiled package with a `pip`
installable package. Also the controller called `udronerc` will receive
much more updates over the next month while the `udrone` package for
drones is fairly finished.

Moving the controller package to git@github.com:aparcar/udronerc.git

Signed-off-by: Paul Spooren <mail@aparcar.org>